fix(camera): Change UIModalPresentationStyle.CurrentContext to FullScreen.
iOS camera does not show in nested <page-router-outlets> or when called from a modal.
As discussed in #109 and #78.

Found that the issue for me was the UIModalPresentationStyle. When mixed with the view hierarchy the camera shows behind the modal.
Might also be worth pushing this as an option as my guess is this may change functionality on iPad, but this is not yet tested.