fix(datetimepicker): spinner & calendar ipads display issue
This commit fixes an issue with the datetimepicker plugin. On iOS iPads
the DatePicker was almost completly cut off in Spinner mode. Only part
of the month was visible so it was unusable. Also, in Calendar mode, the
left and right sides were partially clipped off. This commit fixes those
issues on iOS tablets.
Also, in the demo and demo-angular apps, added an example to show
spinner mode on iOS v14+ bc with 14+ the default is the calendar mode so
the examples only showed calendar mode in iOS 14+.

Fixes NativeScript/plugins#128