fix(camera): Handle 'Limited' iOS Photo permission.
Limited means "we have access to some photos and you can save photos
from the application".

Thus, when we give a limited access to the gallery:
  - succeed
  - save to gallery

In this patch, also add code to check that the switch is exhaustive,
such that compilation will fail and we won't forget to handle future
cases.